refactor : getAnswer isAuthor
